+++ tests/conftest.py
@@ -68,40 +68,6 @@ def stc_language(request):
     return request.param
 
 
-class _LazyPerFunctionEpyccel:
-    """
-    Fallback proxy used when a whole-module `epyccel()` translation fails.
-
-    Compiles each attribute individually and caches it, preserving
-    per-test failure isolation: a single broken function fails only the
-    test(s) that use it, rather than every test sharing the module.
-    """
-    def __init__(self, pymod, language):
-        self._pymod = pymod
-        self._language = language
-        self._cache = {}
-
-    def __getattr__(self, name):
-        if name not in self._cache:
-            print("Translating function")
-            self._cache[name] = epyccel(getattr(self._pymod, name), language=self._language)
-        return self._cache[name]
-
-
-def epyccel_module_with_fallback(pymod, language):
-    """
-    Translate `pymod` as a whole module in one pass.
-
-    If the whole-module translation fails (e.g. `PyccelError` from the
-    translator/compiler, or `ImportError` from the shared-library load
-    check in `epyccel()`), fall back to lazily translating individual
-    functions on first access.
-    """
-    try:
-        return epyccel(pymod, language=language)
-    except Exception:
-        return _LazyPerFunctionEpyccel(pymod, language)
-
 
 @pytest.fixture(autouse=True)
 def skipif_by_language(request):
